requirements.txt
flask==1.1.2
flask-cors==3.0.8
flask-sqlalchemy==2.4.4
flask-jwt-extended==3.24.1
psycopg2==2.8.5
pandas==1.0.5
cryptography==3.0
pillow==7.2.0
